feat: more DBQueryFilterOperators for array fields
not-in, array-contains, array-contains-any
Modeled after Firebase.

There are no tests yet that tests those, but InMemoryDB implements them all already.

There's one "technically breaking change" that '=' Operator is renamed to '=='.
Not marking the release as BREAKING, cause TypeScript should warn on these cases and
should be trivial to adopt. Sorry, but it's too much hassle now to bump majors often.