feat: remove TM from CommonDao
This is a first step in CommonDao refactoring.
TM should not be a concern of db-lib, but instead should be
handled in userland.

Also, `ObjectWithId` is replaced with `Partial<ObjectWithId>` in
CommonDB and CommonDao.
This is to "simplify" types.
It is not exactly correct, but the fact that DBM was "too different" from BM
caused typing issues in the past.
With this change DBM type is more permissive and independent from BM.